Add Alphanumeric class
The common behaviours across different classes for naming purposes
cause code duplication.

Extracting the common bahaviour into an abstract class allows us to use
polymorphism when dealing with names for different purposes. For
example, the error message can be overridden for the Name of a person
and a Title of a document.

Let's pull up bahviours common to all naming attribute classes into a
new abstract class named Alphanumeric. It has the same methods as the
original Name class.
